move check_modes, to match metrics
