+++ specparam/modes/__init__.py
@@ -2,4 +2,4 @@
 
 from .mode import Mode
 from .params import ParamDefinition
-from .definitions import check_modes
+from .check import check_modes
